[PSI] Only send modification events in 'KtCodeFragment' in real IDE
IntelliJ Core does not support file editing, so the colon trembling in
'KtCodeFragment' throws exceptions in unit tests.

(cherry picked from commit 44ef3ed045936cec8c88c92ef52b66154a353f68)

+++ compiler/psi/src/org/jetbrains/kotlin/psi/KtCodeFragment.kt
@@ -16,6 +16,7 @@
 
 package org.jetbrains.kotlin.psi
 
+import com.intellij.mock.MockProject
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Key
@@ -131,14 +132,18 @@ abstract class KtCodeFragment(
     }
 
     override fun addImportsFromString(imports: String?) {
+        val notifyChanged = viewProvider.isEventSystemEnabled && project !is MockProject
+
         if (imports != null && appendImports(imports)) {
-            // For K1: This forces the code fragment to be re-highlighted.
-            add(KtPsiFactory(project).createColon()).delete()
+            if (notifyChanged) {
+                // This forces the code fragment to be re-highlighted
+                add(KtPsiFactory(project).createColon()).delete()
+            }
 
             // Increment the modification stamp
             clearCaches()
 
-            if (viewProvider.isEventSystemEnabled) {
+            if (notifyChanged) {
                 project.messageBus.syncPublisher(IMPORT_MODIFICATION).onCodeFragmentImportsModification(this)
             }
         }
